streamlit_app/requirements.txt
mysql-connector-python
streamlit
streamlit-option-menu
streamlit_authenticator
python-dotenv